moved images to webpack for production build
