+++ client/src/components/Landing.js
@@ -5,7 +5,7 @@ import { connect } from 'react-redux';
 import { fetchCurrentUser } from '../actions/index';
 import { Button } from '@material-ui/core';
 import '../style/dashboard.css';
-import logo from '../Images/KnowMe.png';
+import logo from '../static/Images/KnowMe.png';
 
 class Landing extends Component {
   componentDidMount() {
