+++ client/src/static/Images/question.png
