+++ client/src/style/style.css
@@ -2,7 +2,7 @@
 body {
   width: 100%;
   height: 100%;
-  background-image: url(../Images/bg-clouds.png);
+  background-image: url(../static/Images/bg-clouds.png);
   background-position: 0px 0px;
   background-repeat: repeat-x;
   background-size: cover;
